# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Social Media API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "servers": [
    {
      "url": "http://localhost:3000/"
    },
    {
      "url": "https://localhost:3000/"
    }
  ],
...
  "paths": {
    "/api/users/": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "To get all the users in the given page and size.",
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "description": "the page user is in",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "size",
            "in": "query",
            "description": "the number of entities in each page.",
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "description": "the parameter that entities will be sorted by. options: [createdAt]. for descending sort add - before parmeter.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "an array of users",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "post": {
        "tags": [
          "User"
        ],
        "description": "to create a new user.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "the new user object + user token in header with x-auth-token key.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddUser"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/AddUser"
              }
            }
          }
        }
      }
    },
    "/api/users/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "description": "to get a user data by given id.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "id of the user."
          }
        ],
        "responses": {
          "200": {
            "description": "the user object",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "put": {
        "tags": [
          "User"
        ],
        "description": "to update a user by id and the user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "id of the user."
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "the new user object + the new user token in header with x-auth-token key.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddUser"
              }
            },
            "application/xml": {
              "schema": {
                "$ref": "#/components/schemas/AddUser"
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "User"
        ],
        "description": "to delete a user data by given id and a the user",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "id of the user."
          },
          {
            "name": "x-auth-token",
            "in": "header",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "the user object",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              },
              "application/xml": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/User"
                  },
                  "xml": {
                    "name": "main"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Not Found"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/users/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "to login a user.",
        "parameters": [
          {
            "name": "emailOrName",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "emailOrName": {
                  "example": "any"
                },
                "password": {
                  "example": "any"
                }
              }
            },
            "description": "the username or email of the user.",
            "required": true
          },
          {
            "name": "password",
            "in": "body",
            "description": "the password of the user.",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {}
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "emailOrName": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ali"
          },
          "username": {
            "type": "string",
            "example": "username1"
          },
          "age": {
            "type": "number",
            "example": 29
          },
          "email": {
            "type": "string",
            "example": "felan@felan.com"
          },
          "phoneNumber": {
            "type": "string",
            "example": "01234567890"
          },
          "password": {
            "type": "string",
            "example": "password"
          },
          "posts": {
            "type": "array",
            "items": {
              "xml": {
                "name": "post"
              },
              "$ref": "#/components/schemas/Post"
            }
          }
        },
        "required": [
          "name",
          "username",
          "email",
          "phoneNumber",
          "password"
        ],
        "xml": {
          "name": "User"
        }
      },
      "Post": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "example": "title"
          },
          "text": {
            "type": "string",
            "example": "sample text for the post..."
          },
          "author": {
            "xml": {
              "name": "user"
            },
            "$ref": "#/components/schemas/User"
          },
          "likes": {
            "type": "array",
            "items": {
              "xml": {
                "name": "like"
              },
              "$ref": "#/components/schemas/Like"
            }
          }
        },
        "required": [
          "title",
          "text",
          "author"
        ],
        "xml": {
          "name": "Post"
        }
      },
      "Like": {
        "type": "object",
        "properties": {
          "postId": {
            "xml": {
              "name": "post"
            },
            "$ref": "#/components/schemas/Post"
          },
          "userId": {
            "xml": {
              "name": "user"
            },
            "$ref": "#/components/schemas/User"
          }
        },
        "required": [
          "postId",
          "userId"
        ],
        "xml": {
          "name": "Like"
        }
      },
      "AddUser": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "example": "ali"
          },
          "username": {
            "type": "string",
            "example": "username1"
          },
          "age": {
            "type": "number",
            "example": 29
          },
          "email": {
            "type": "string",
            "example": "felan@felan.com"
          },
          "phoneNumber": {
            "type": "string",
            "example": "01234567890"
          },
          "password": {
            "type": "string",
            "example": "password"
          }
        },
        "required": [
          "name",
          "username",
          "email",
          "phoneNumber",
          "password"
        ],
        "xml": {
          "name": "AddUser"
        }
      },
...
    }
  }
}